+++ apps/server-auth/src/helpers/hashPassword.ts
@@ -1,0 +1,3 @@
+import { hashSync } from "bcrypt";
+
+export const hashPassword = (password: string) => hashSync(password, 12);