+++ apps/server-auth/src/services/authZServices.ts
@@ -1,0 +1,15 @@
+import { AuthDocument } from "models/Auth/types";
+
+export const suspendClient = (
+  authClient: AuthDocument,
+  suspention: {
+    suspentionLiftTime: Date;
+    suspentionReason: string;
+  }
+) => {
+  authClient.isSuspended = true;
+  authClient.suspentionLiftTime = suspention.suspentionLiftTime;
+  authClient.suspentionReason = suspention.suspentionReason;
+
+  return authClient.save();
+};