+++ packages/shared-types/IPaginatedResult.ts
@@ -1,0 +1,8 @@
+export interface IPaginatedResult<T> {
+  items: T[];
+  page: string;
+  hasNextPage: boolean;
+  hasPreviousPage: boolean;
+  totalItems: number;
+  totalPages: number;
+}